requirements.txt
# Non Python Deps
# sudo apt  install python-dev python-pip libxml2-dev libxslt1-dev zlib1g-dev libffi-dev libssl-dev
certifi>=2020.6.20
chardet>=3.0.4
flake8>=3.8.4
idna>=2.10
mccabe>=0.6.1
pycodestyle>=2.6.0
pyflakes>=2.2.0
PyYAML>=5.3.1
requests>=2.24.0
urllib3>=1.25.11
pyOpenSSL>=19.1.0
pushbullet.py>=0.8.1